feat: add status codes

+++ app/Http/Controllers/MarcaController.php
@@ -17,7 +17,7 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas, 200);
     }
 
     /**
@@ -34,7 +34,7 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -43,7 +43,12 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
+        return response()->json($marca, 200);
     }
 
     /**
@@ -60,8 +65,13 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização, o recurso solicitado não existe'], 404);
+        }
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -70,7 +80,12 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão, o recurso solicitado não existe'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi removida com sucesso'];
+        return response()->json(['msg' => 'A marca foi removida com sucesso'], 200);
     }
 }
